move outlets to new lib layout
