+++ projects/storefrontlib/src/cms-components/user/logout/logout.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
-import { LogoutGuard } from './logout-guard';
 import { RouterModule } from '@angular/router';
-import { PageLayoutComponent } from '../../../lib/cms/index';
+import { PageLayoutComponent } from '../../../cms-structure/page/index';
+import { LogoutGuard } from './logout-guard';
 
 @NgModule({
   imports: [
