+++ projects/storefrontlib/src/cms-structure/index.ts
@@ -1,2 +1,3 @@
+export * from './outlet/index';
 export * from './page/index';
 export * from './seo/index';
