+++ projects/storefrontlib/src/cms-structure/outlet/outlet-style.service.ts
