+++ projects/storefrontlib/src/cms-structure/outlet/outlet.model.ts
