+++ projects/storefrontlib/src/cms-structure/page/index.ts
@@ -1,13 +1,9 @@
-// model
-export * from './model/index';
-
-// slot
-export * from './slot/page-slot.module';
-export * from './slot/page-slot.component';
-
-// component
-export * from './component/page-component.module';
 export * from './component/component-wrapper.directive';
-
-// configs
+export * from './component/page-component.module';
 export * from './config/default-cms-content.config';
+export * from './model/index';
+export * from './page-layout/page-layout.component';
+export * from './page-layout/page-layout.module';
+export * from './page-layout/page-layout.service';
+export * from './slot/page-slot.component';
+export * from './slot/page-slot.module';
