+++ projects/storefrontlib/src/cms-structure/page/page-layout/page-layout.module.ts
@@ -1,12 +1,12 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { PageLayoutComponent } from './page-layout.component';
-import { CmsModule } from '../cms.module';
-import { PageLayoutService } from './page-layout.service';
+import { NgModule } from '@angular/core';
 import { PageSlotModule } from '../../../cms-structure/page/slot/page-slot.module';
+import { OutletModule } from '../../outlet/outlet.module';
+import { PageLayoutComponent } from './page-layout.component';
+import { PageLayoutService } from './page-layout.service';
 
 @NgModule({
-  imports: [CommonModule, CmsModule, PageSlotModule],
+  imports: [CommonModule, OutletModule, PageSlotModule],
   declarations: [PageLayoutComponent],
   providers: [PageLayoutService],
   exports: [PageLayoutComponent],
