+++ projects/storefrontlib/src/layout/layout.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Config, ConfigModule } from '@spartacus/core';
-import { OutletRefModule } from '../lib/outlet/index';
-import { StyleRefModule } from '../lib/outlet/style-ref/style-ref.module';
+import { OutletRefModule } from '../cms-structure/outlet/index';
+import { StyleRefModule } from '../cms-structure/outlet/style-ref/style-ref.module';
 import { defaultLayoutConfig } from './config/default-layout-config';
 import { LayoutConfig } from './config/layout-config';
 import { MainModule } from './main/main.module';
