+++ projects/storefrontlib/src/lib/cms/index.ts
@@ -1,6 +1,4 @@
 export * from './cms.module';
 export * from './guards/cms-page.guard';
-export * from './page-layout/page-layout.module';
-export * from './page-layout/page-layout.component';
 export * from './services/cms-mapping.service';
 export * from './services/cms-routes.service';
