--- projects/storefrontlib/src/lib/occ/index.ts
@@ -1,1 +1,0 @@
-export * from './occ.module';
