+++ projects/storefrontlib/src/lib/suffix-routes/suffix-routes.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { PageLayoutComponent } from '../../cms-structure/page/index';
 import { CmsPageGuard } from '../cms/guards/cms-page.guard';
-import { PageLayoutComponent } from '../cms/page-layout/page-layout.component';
-import { suffixUrlMatcher } from './suffix-url-matcher';
 import { ProductPageComponent } from '../ui/pages/product-page/product-page.component';
+import { suffixUrlMatcher } from './suffix-url-matcher';
 
 @NgModule({
   imports: [
