+++ projects/storefrontlib/src/lib/ui/pages/order-confirmation-page/order-confirmation-page.module.ts
@@ -1,15 +1,12 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-
 import { AuthGuard } from '@spartacus/core';
-
-import { OrderConfirmationModule } from '../../../checkout/index';
-import { PageLayoutModule } from '../../../cms/index';
-import { OutletRefModule } from '../../../outlet/index';
+import { OutletRefModule } from '../../../../cms-structure/outlet/index';
+import { PageLayoutModule } from '../../../../cms-structure/page/page-layout/page-layout.module';
 import { OrderConfirmationPageGuard } from '../../../checkout/guards/order-confirmation-page.guard';
+import { OrderConfirmationModule } from '../../../checkout/index';
 import { CmsPageGuard } from '../../../cms/guards/cms-page.guard';
-
 import { OrderConfirmationPageComponent } from './order-confirmation-page.component';
 
 const routes: Routes = [
