+++ projects/storefrontlib/src/public_api.ts
@@ -5,13 +5,13 @@
 export * from './cms-components/index';
 export * from './cms-components/myaccount/index';
 export * from './cms-structure/index';
+export * from './cms-structure/outlet/index';
 export * from './layout/index';
 export * from './lib/checkout/index';
 export * from './lib/cms-lib/index';
 export * from './lib/cms-route/index';
 export * from './lib/cms/index';
 export * from './lib/occ/index';
-export * from './lib/outlet/index';
 export * from './lib/pwa/pwa.module';
 export * from './lib/store-finder/index';
 export * from './lib/storefront-config';
